# Testing/Clase_2_1/docs/swagger.yaml
definitions:
  handler.request:
    properties:
      Receptor:
        type: string
      amount:
        type: number
      code:
        type: string
      coin:
        type: string
      emitting:
        type: string
      id:
        type: integer
      transactionDate:
        type: string
    type: object
  web.Response:
    properties:
      code:
        type: string
      data: {}
      error:
        type: string
    type: object
info:
  contact:
    name: API Support
    url: https://developers.mercadolibre.com.ar/support
  description: this API handler MELI transactions
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://developers.mercadolibre.com.ar/es_ar/terminos-y-condiciones
  title: MELI BootCamp API
  version: "1.0"
paths:
  /transactions:
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List transactions
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.Response'
        "404":
          description: Not found products
          schema:
            $ref: '#/definitions/web.Response'
        "500":
          description: 'Internal server error '
          schema:
            $ref: '#/definitions/web.Response'
      summary: Show list transactions
      tags:
      - Transactions
    post:
      consumes:
      - application/json
      parameters:
      - description: token requerido
        in: header
        name: token
        required: true
        type: string
      - description: Transaction to Store
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.Response'
        "404":
          description: Not found products
          schema:
            $ref: '#/definitions/web.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.Response'
      summary: Store transaction
      tags:
      - Transactions
  /transactions/{Id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Id transactions
        in: path
        name: Id
        required: true
        type: integer
      - description: token
        in: header
        name: token
        type: string
      - description: transaction to update
        in: body
        name: transactions
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.Response'
      summary: Update transaction
      tags:
      - Transactions
  /transactions/{id}:
    delete:
      parameters:
      - description: Transaction id
        in: path
        name: id
        required: true
        type: integer
      - description: Token
        in: header
        name: token
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.Response'
      summary: Delete transaction
      tags:
      - Transactions
    patch:
      consumes:
      - application/json
      description: This endpoint update field code and amount from an transaction
      parameters:
      - description: Token header
        in: header
        name: token
        required: true
        type: string
      - description: Transaction Id
        in: path
        name: Id
        required: true
        type: integer
      - description: transaction to update
        in: body
        name: transactions
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.Response'
      summary: Update code and amount transactions
      tags:
      - Transactions
swagger: "2.0"